Remove static tx validation from RPC
